# doc/api.yaml
openapi: 3.0.3
servers:
  - url: "http://localhost:3000"
info:
  title: WasaText
  description: |
    Connect with your friends effortlessly using WASAText! Send and receive messages, whether one-on-one
    or in groups, all from the convenience of your PC. Enjoy seamless conversations with text or GIFs and
    easily stay in touch through your private chats or group discussions.
  version: 0.0.1
security:
  - bearerAuth: []
...
paths:
...
  /users/{user}/groups:
    parameters:
      - $ref: "#/components/parameters/uid"
    post:
      tags: [Group]
      operationId: createGroup
      summary: Create group
      description: Creates a new group.
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GroupInput" }
        required: true
      responses:
        "201":
          description: Group created successfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Group" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/ServerError" }
  /users/{user}/groups/{group}:
    parameters:
      - $ref: "#/components/parameters/uid"
      - $ref: "#/components/parameters/gid"
    put:
      tags: [Group]
      operationId: addToGroup
      summary: Add member
      description: Adds user to group.
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Users" }
        required: true
      responses:
        "200":
          description: The user was added to the group
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/Success" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/ServerError" }
    delete:
      tags: [Group]
      operationId: leaveGroup
      summary: Leave group
      description: Removes user from group.
      responses:
        "200":
          description: The user has left the group
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/Success" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "500": { $ref: "#/components/responses/ServerError" }
  /users/{user}/groups/{group}/groupname:
    parameters:
      - $ref: "#/components/parameters/uid"
      - $ref: "#/components/parameters/gid"
    put:
      tags: [Group]
      operationId: setGroupName
      summary: Set group name
      description: Updates group name.
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GroupNameInput" }
        required: true
      responses:
        "200":
          description: The group name was set successfully
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/Success" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /users/{user}/groups/{group}/groupphoto:
    parameters:
      - $ref: "#/components/parameters/uid"
      - $ref: "#/components/parameters/gid"
    put:
      tags: [Group]
      operationId: setGroupPhoto
      summary: Set group photo
      description: Updates group photo. Send an empty string to remove the photo.
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PhotoInput" }
        required: true
      responses:
        "200":
          description: The photo was uploaded successfully
          content:
            plain/text:
              schema: { $ref: "#/components/schemas/Success" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }

# Reusable components (schemas, parameters, responses)
components:
...
  parameters:
    uid:
      name: user
      in: path
      required: true
      description: User ID
      schema: { $ref: "#/components/schemas/ResourceId" }
    gid:
      name: group
      in: path
      required: true
      description: Group ID
      schema: { $ref: "#/components/schemas/ResourceId" }
...
  schemas:
    Success:
      title: Successful
      description: The operation was successful
      type: string
      pattern: "^.*?$"
      example: "Action successfully completed"
      minLength: 2
      maxLength: 30
    # Elements
    ResourceId:
      title: ID - identification
      description: General identification number (used for example for users, groups or conversation)
      type: integer
      example: 100
      minimum: 1
    Photo:
      title: Photo
      description: Photo of a user profile or a group
      type: string
      format: byte
      minLength: 0
      maxLength: 9999
    Name:
      title: Name
      description: Name of a user or a group
      type: string
      pattern: ^[a-zA-Z0-9 _-]{3,30}$
      example: "user123"
      minLength: 3
      maxLength: 13
...
    User:
      title: User profile
      description: User profile information
      type: object
      properties:
        username: { $ref: "#/components/schemas/Name" }
        photo: { $ref: "#/components/schemas/Photo" }
        user_id: { $ref: "#/components/schemas/ResourceId" }
      required: [username, user_id]
      example:
        username: "MarioRossi"
        user_id: 10
    Users:
      title: List of profile
      description: List of users in a group, conversation or in the search operation
      type: object
      properties:
        users:
          type: array
          description: All users returned after a request like search
          items: { $ref: "#/components/schemas/User" }
          uniqueItems: true
          minItems: 0
          maxItems: 100
      required: [users]
      example:
        users:
          - username: "MarioRossi"
            user_id: 10
...
    Group:
      title: Group
      description: Group in which various users write each other
      type: object
      properties:
        groupname: { $ref: "#/components/schemas/Name" }
        photo: { $ref: "#/components/schemas/Photo" }
        group_id: { $ref: "#/components/schemas/ResourceId" }
        members: { $ref: "#/components/schemas/Users" }
      required: [groupname, group_id, members]
      example:
        groupname: "WasaGroup"
        group_id: 42
        members:
          users:
            - username: "MarioRossi"
              user_id: 10
            - username: "LuigiVerdi"
              user_id: 11
...
    PhotoInput:
      description: Photo to upload in the profile or group
      type: object
      properties:
        photo: { $ref: "#/components/schemas/Photo" }
      required:
        - photo
    GroupInput:
      description: Name for the new group
      type: object
      properties:
        name: { $ref: "#/components/schemas/Name" }
        members: { $ref: "#/components/schemas/Users" }
      required:
        - name
        - members
    GroupNameInput:
      description: New name for the group
      type: object
      properties:
        name: { $ref: "#/components/schemas/Name" }
      required:
        - name
    # Used for the Error responses (40x and 50x)
    ErrorItem:
      type: object
      description: Description of the error
      properties:
        code:
          description: Code error
          type: string
          example: "not_found"
          enum: ["bad_request", "unauthorized", "forbidden", "not_found", "conflict", "payload_too_large", "server_error"]
        reason:
          type: string
          description: Explanation of the reason for the error which can be shown to a client user
          pattern: "^.*?$"
          example: Page not found
          minLength: 1
          maxLength: 100
        message:
          type: string
          description: More details of the error
          pattern: "^.*?$"
          example: "The requested URL was not found"
          minLength: 1
          maxLength: 100
        more_info:
          type: string
          description: URL or text with more info
      required:
        - code
        - reason
    Error:
      description: Object describe an error with the code and the message
      type: object
      properties:
        trace:
          type: string
          description: Error trace id
        errors:
          description: Array of error codes generated by the operation
          type: array
          items: { $ref: "#/components/schemas/ErrorItem" }
          minItems: 1
          maxItems: 100
      required:
        - errors
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: The access token is missing or it's expired
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: The user is not allowed to perform the action
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ServerError:
      description: The server encountered an internal error. Further info in server logs
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    PayloadTooLarge:
      description: The request entity is larger than limits defined by server
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }